+++ src/main/java/hellfirepvp/astralsorcery/client/screen/container/ScreenContainerAltarRadiance.java
@@ -128,7 +128,7 @@ public class ScreenContainerAltarRadiance extends ScreenContainerAltar<Container
             spriteStarlight.getResource().bindTexture();
 
             int tick = altar.getTicksExisted();
-            Tuple<Double, Double> uvOffset = spriteStarlight.getUVOffset(tick);
+            Tuple<Float, Float> uvOffset = spriteStarlight.getUVOffset(tick);
             drawRect(guiLeft + 11, guiTop + 104, (int) (232 * percFilled), 10,
                     uvOffset.getA(), uvOffset.getB(),
                     spriteStarlight.getULength() * percFilled, spriteStarlight.getVLength());
