+++ src/main/java/hellfirepvp/astralsorcery/common/perk/tree/PerkTreeGem.java
@@ -92,7 +92,7 @@ public class PerkTreeGem<T extends AbstractPerk & GemSlotPerk> extends PerkTreeP
 
         double uLength = tex.getULength();
         double vLength = tex.getVLength();
-        Tuple<Double, Double> frameUV = tex.getUVOffset(spriteOffsetTick);
+        Tuple<Float, Float> frameUV = tex.getUVOffset(spriteOffsetTick);
 
         for (int i = 0; i < 4; i++) {
             int u = ((i + 1) & 2) >> 1;
